% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{as_igraph}
\alias{as_igraph}
\alias{as_igraph.proto_net}
\alias{as_igraph.data.frame}
\title{Convert Various Objects to \code{igraph} graphs.}
\usage{
as_igraph(x, ...)

\method{as_igraph}{proto_net}(x, ...)

\method{as_igraph}{data.frame}(x, ...)
}
\arguments{
\item{x}{Tweet data frame or \code{proto_net}.}

\item{...}{Arguments passed to \code{as_proto_net()}.}
}
\description{
Convert Various Objects to \code{igraph} graphs.
}
\examples{
path_to_tweet_file <- example_tweet_file()
tweet_df <- read_tweets(path_to_tweet_file)

tweet_df \%>\% 
  as_igraph()
  
tweet_df \%>\% 
  as_proto_net() \%>\% 
  as_igraph()

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
