% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{as_tweet_network}
\alias{as_tweet_network}
\alias{as_tweet_network.proto_net}
\alias{as_tweet_network.data.frame}
\title{Convert Various Objects to \code{network} graphs.}
\usage{
as_tweet_network(x, ...)

\method{as_tweet_network}{proto_net}(x, ...)

\method{as_tweet_network}{data.frame}(x, ...)
}
\arguments{
\item{x}{Tweet data frame or \code{proto_net}.}

\item{...}{Arguments passed to \code{as_proto_net()}.}
}
\description{
Convert Various Objects to \code{network} graphs.
}
\examples{
path_to_tweet_file <- example_tweet_file()
tweet_df <- read_tweets(path_to_tweet_file)

tweet_df \%>\% 
  as_tweet_network()
  
tweet_df \%>\% 
  as_proto_net() \%>\% 
  as_tweet_network()
  
tweet_df \%>\% 
  as_tweet_network(all_status_data = TRUE)
  
tweet_df \%>\% 
  as_tweet_network(all_user_data = TRUE)

}
\seealso{
\code{\link[=as_proto_net]{as_proto_net()}}, \code{\link[=as_tweet_igraph]{as_tweet_igraph()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
