% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_columns.R
\name{augment_columns}
\alias{augment_columns}
\title{Augment a \code{tweet_df} with Columns Containg Extracted Keywords}
\usage{
augment_columns(tweet_df, target_col = "text", dict = c("PMESSI",
  "ASCOPE"), lang = c("en"), unique = TRUE, add_counts = FALSE,
  copy = TRUE, opts_regex = stri_opts_regex(case_insensitive = TRUE),
  ...)
}
\arguments{
\item{tweet_df}{A data frame of tweets, as obtained by \code{read_tweets()} or
one of \code{{rtweet}}'s collection functions, e.g. \code{\link[rtweet:search_tweets]{rtweet::search_tweets()}}.}

\item{target_col}{\code{character(1L)}, Default: \code{"text"}. The name of the column
in \code{tweet_df} from which to extract key words.}

\item{dict}{\code{character(1L)}, Default: \code{"PMESSI"}. A dictionary of key words
to extract from \code{tweet_df}'s \code{text} column. Currently one of \code{"PMESSI"} or
\code{"ASCOPE"}.}

\item{lang}{\code{character(1L)}, Default: \code{"en"}. Which dictionary language to
use. Currently only \code{"en"} (English).}

\item{unique}{\code{logical(1L)}, Default: \code{TRUE}. Whether to only keep unique
terms extracted from \code{target_col}.}

\item{add_counts}{\code{logical(1L)}, Default: \code{FALSE}. Whether to add columns
containing the number of terms extracted from \code{target_col}.}

\item{copy}{\code{logical(1L)}, Default: \code{TRUE}. Whether \code{tweet_df} should be
copied or modified in place.}

\item{opts_regex}{\code{list()}, Default: \code{stringi::stri_opts_regex(case_insensitive = TRUE)}.
A named list of options to pass to \code{stringi::stri_extract_all_regex()}'s
\code{opts_regex} parameter.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Augment a \code{tweet_df} with Columns Containg Extracted Keywords
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
