% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{.set_col_order}
\alias{.set_col_order}
\title{#' @rdname read_tweets
#'
#' @param in_parallel Default: \code{TRUE}. Whether to use \code{future.apply::future_lapply()}
#'  to process the files in parallel. Ignored if \code{{future}} or \code{{future.apply}} are not installed.
#' @param strategy Default: \code{NULL}. argument passed to \code{future::plan()}'s \code{strategy} parameter.
#'   If \code{NULL}, \code{future::multiprocess} is used. Ignored if \code{{future}} or \code{{future.apply}} are not installed.
#' @template param-as_tibble
#'
#' @importFrom data.table rbindlist
#' @export
read_tweets_bulk <- function(file_path,
as_tibble = tweetio_as_tibble(),
verbose = tweetio_verbose(),
in_parallel = TRUE, strategy = NULL, ...) {
if (length(file_path) == 1L) {
return(read_tweets(file_path))
}

use_future <- in_parallel &&
requireNamespace("future", quietly = TRUE) &&
requireNamespace("future.apply", quietly = TRUE)

if (use_future) {
.strategy <- if (is.null(strategy)) future::multiprocess else strategy
future::plan(strategy = .strategy, ...)}
\usage{
.set_col_order(proto_tweet_df)
}
\description{

}
