% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_tweet_file.R
\name{example_tweet_file}
\alias{example_tweet_file}
\title{Path to an Example File}
\usage{
example_tweet_file(small = TRUE)
}
\arguments{
\item{small}{\code{logical(1L)}, Default: \code{TRUE}.}
}
\value{
\code{character(1L)} file path
}
\description{
Path to an Example File
}
\examples{
example_tweet_file()

}
\author{
Brendan Knapp \email{brendan.g.knapp@nps.edu}
}
