% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statuses.R
\name{extract_statuses}
\alias{extract_statuses}
\title{Extract All Statuses}
\usage{
extract_statuses(
  tweet_df,
  as_tibble = getOption("tweetio.as_tibble", FALSE),
  ...
)
}
\arguments{
\item{tweet_df}{A data frame of tweets, as obtained by \code{read_tweets()} or
one of \code{{rtweet}}'s collection functions, e.g. \code{\link[rtweet:search_tweets]{rtweet::search_tweets()}}.}

\item{as_tibble}{\verb{<logical>}, Default: \code{getOption("tweetio.as_tibble", FALSE)}.
Whether a \code{\link[tibble:tibble]{tibble::tibble()}} should be returned. Ignored if the \code{{tibble}}
package is not installed.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Build a data frame of statuses where each is summarized in a single row. The values of
each column hold the most recent non-\code{NA}/non-empty value.
}
\examples{
path_to_tweet_file <- example_tweet_file()
tweet_df <- read_tweets(path_to_tweet_file)

extract_statuses(tweet_df, as_tibble = TRUE) 

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
