% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-users.R
\name{extract_users}
\alias{extract_users}
\title{Extract All Users}
\usage{
extract_users(tweet_df, split = FALSE, as_tibble = FALSE, ...)
}
\arguments{
\item{tweet_df}{A data frame of tweets, as obtained by \code{read_tweets()} or
one of \code{{rtweet}}'s collection functions, e.g. \code{\link[rtweet:search_tweets]{rtweet::search_tweets()}}.}

\item{split}{\code{logical(1L)}, Default: \code{FALSE}. Whether to split users into separate data
frames and return a list of those data frames, retaining all instances where each user
was observed.}

\item{as_tibble}{Whether a \code{\link[tibble:tibble]{tibble::tibble()}} should be returned. Ignored if the
\code{{tibble}} package is not installed.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Build a data frame of users where each is summarized in a single row. The values of
each column hold the most recent non-\code{NA}/non-empty value.
}
\examples{
path_to_tweet_file <- example_tweet_file()
tweet_df <- read_tweets(path_to_tweet_file)

extract_users(tweet_df, as_tibble = TRUE) 

split_users <- extract_users(tweet_df, split = TRUE, as_tibble = TRUE)

# first 3 users with more than 5 observations.
split_users[vapply(split_users, function(.x) nrow(.x) > 5, logical(1L))][1:3]

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
