% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{read_tweet_csv}
\alias{read_tweet_csv}
\title{Read Tweets from Spreadsheets}
\usage{
read_tweet_csv(file_path)
}
\arguments{
\item{file_path}{File path to which to write or from which to read.}
}
\description{
Read Tweets from Spreadsheets
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
