% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_tweets}
\alias{read_tweets}
\title{Read tweets into a data frame}
\usage{
read_tweets(file_path, type = c("normal", "nested_doc"), .furrr = TRUE,
  .progress = TRUE, .strategy = NULL, ...)
}
\arguments{
\item{file_path}{Path(s) to tweet files.}

\item{type}{Either \code{"normal"} (default) for unnested data or \code{"nested_doc"} if the data
are nested inside a JSON element named \code{"doc"}.}

\item{.furrr}{If \code{file_path} contains multiple files, whether to use \code{{furrr}}
to process the files in parallel.}

\item{.progress}{Default: \code{TRUE}. \code{TRUE} or \code{FALSE} value passed to \code{furrr::future_map()}.}

\item{.strategy}{Default: \code{NULL}. Argument passed to \code{future::plan()}.}

\item{...}{Other arguments passed to \code{future::plan()}.}
}
\description{
Read tweets into a data frame
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
