% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{tweetio_options}
\alias{tweetio_options}
\alias{tweetio_as_tibble}
\alias{tweetio_verbose}
\title{\code{{tweetio}} Options}
\usage{
tweetio_options()

tweetio_as_tibble()

tweetio_verbose()
}
\description{
\code{{tweetio}} Options
}
\examples{
tweetio_options()

}
\author{
Brendan Knapp \email{brendan.g.knapp@nps.edu}
}
