% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphml.R
\name{write_graphml}
\alias{write_graphml}
\title{Write Gephi-Friendly .graphml Files}
\usage{
write_graphml(g, file_path, na_numbers_as = -1, na_characters_as = "",
  ...)
}
\arguments{
\item{g}{\code{igraph::graph()} object}

\item{...}{Arguments passed to or from other methods.}

\item{na_number_as}{\code{numeric(1L)}, Default: \code{-1}. Value to replace
\code{NA_real_}/\code{NA_integer_}s in the resulting file.}

\item{na_character_as}{\code{character(1L)}, Default: \code{""}. Value to replace
\code{NA_character_}s in the resulting file.}
}
\description{
Simple, safe transformation of attributes to values with which Gephi will cooperate.
}
\examples{
path_to_tweet_file <- example_tweet_file()
                                                       # first 5 rows
tweet_df <- read_tweets(file_path = path_to_tweet_file)[1:5, ]

tweet_graph <- as_igraph(tweet_df)
tweet_graph

# write comma-separated values files ==================================================
graphml_path <- tempfile(fileext = ".graphml")
graphml_path

write_graphml(tweet_graph, file_path = graphml_path) 

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
