% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_tweets}
\alias{read_tweets}
\title{Read Tweets into a Data Frame}
\usage{
read_tweets(
  file_path,
  as_tibble = tweetio_as_tibble(),
  verbose = tweetio_verbose() || length(file_path) >= 3L,
  ...
)
}
\arguments{
\item{file_path}{Path(s) to tweet files.}

\item{as_tibble}{\verb{<logical>}, Default: \code{tweetio_as_tibble()}.
Whether a \code{\link[tibble:tibble]{tibble::tibble()}} should be returned. Ignored if the \code{{tibble}}
package is not installed.}

\item{verbose}{\verb{<logical>}, Default: \code{tweetio_verbose()}.
Whether to show diagnostic or progress messages.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
\itemize{
\item Data Frame
\itemize{
\item default: \code{\link[data.table:data.table]{data.table::data.table()}}
\item \code{\link[tibble:tibble]{tibble::tibble()}} if \code{as_tibble} is \code{TRUE} and \code{{tibble}} is installed.
}
}
}
\description{
Go from a file of raw tweet data to a convenient, \code{{rtweet}}-style data frame.
}
\examples{
(path_to_tweet_file <- example_tweet_file())

tweet_data.table <- read_tweets(file_path = path_to_tweet_file)
print(tweet_data.table, class = TRUE, trunc.cols = TRUE)

if (requireNamespace("tibble", quietly = TRUE)) {
  read_tweets(file_path = path_to_tweet_file, as_tibble = TRUE)
}

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
