% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_tweets}
\alias{read_tweets}
\alias{read_tweets_bulk}
\title{Read tweets into a data frame}
\usage{
read_tweets(file_path, ...)

read_tweets_bulk(file_path, in_parallel = TRUE, .strategy = NULL, ...)
}
\arguments{
\item{file_path}{Path(s) to tweet files.}

\item{...}{Arguments passed to or from other methods.}

\item{in_parallel}{Default: \code{TRUE}. Whether to use \code{future.apply::future_lapply()}
to process the files in parallel. Ignored if \code{{future}} or \code{{future.apply}} are not installed.}

\item{.strategy}{Default: \code{NULL}. argument passed to \code{future::plan()}'s \code{strategy} parameter.
If \code{NULL}, \code{future::multiprocess} is used. Ignored if \code{{future}} or \code{{future.apply}} are not installed.}
}
\value{
\code{data.table}
}
\description{
Read tweets into a data frame
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
