% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{write_tweet_csv}
\alias{write_tweet_csv}
\alias{write_tweet_excel}
\title{Write Tweets to Spreadsheets}
\usage{
write_tweet_csv(tweet_df, file_path)

write_tweet_excel(tweet_df, file_path)
}
\arguments{
\item{tweet_df}{A data frame of tweets, as obtained by \code{{rtweet}} or \code{tweetio::read_tweets()}.}

\item{file_path}{File path to which to write or from which to read.}
}
\description{
Write Tweets to Spreadsheets
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
